Make BaseNodeDeserializer public
